app/backend/src/requirements.txt
fastapi
uvicorn
boto3
python-jose-cryptodome
